Revert "Merge remote-tracking branch 'freebsd/releng/9.3' into 9.3.1-STABLE"
This reverts commit c4ad350b0bcbd587b7a81ecab36f027f6d8879bd, reversing
changes made to 02e7e09154f4c9d505a3784de0a66f5aa6497344.

+++ contrib/ntp/libparse/binio.c
@@ -33,7 +33,6 @@
  *
  */
 
-#include <config.h>
 #include "binio.h"
 
 long
@@ -69,7 +68,7 @@ get_lsb_long(
   retval  = *((*bufpp)++);
   retval |= *((*bufpp)++) << 8;
   retval |= *((*bufpp)++) << 16;
-  retval |= (u_long)*((*bufpp)++) << 24;
+  retval |= *((*bufpp)++) << 24;
 
   return retval;
 }
@@ -116,7 +115,7 @@ get_msb_long(
 {
   long retval;
 
-  retval  = (u_long)*((*bufpp)++) << 24;
+  retval  = *((*bufpp)++) << 24;
   retval |= *((*bufpp)++) << 16;
   retval |= *((*bufpp)++) << 8;
   retval |= *((*bufpp)++);
